Issue #406: Handle events with no vertical
